src/process/requirements.txt
python-dotenv==0.19.2
psycopg2-binary==2.9.1
pandas==1.3.4
python-csv==0.0.13